Extract out ni based collision blocks to bound module.
